refactor: extract trap check task and move start trigger to GameManager
