+++ core/src/main/java/cc/azuramc/bedwars/event/impl/GameStartEvent.java
@@ -5,23 +5,13 @@ import cc.azuramc.bedwars.config.object.EventConfig;
 import cc.azuramc.bedwars.config.object.MessageConfig;
 import cc.azuramc.bedwars.event.AbstractGameEvent;
 import cc.azuramc.bedwars.game.GameManager;
-import cc.azuramc.bedwars.game.GamePlayer;
-import cc.azuramc.bedwars.game.GameTeam;
 import cc.azuramc.bedwars.game.spectator.task.SpectatorCompassTask;
 import cc.azuramc.bedwars.game.task.GeneratorTask;
-import cc.azuramc.bedwars.upgrade.trap.TrapManager;
-import cc.azuramc.bedwars.upgrade.trap.TrapType;
-import cc.azuramc.bedwars.util.LoggerUtil;
-import com.cryptomorin.xseries.XPotion;
 import com.cryptomorin.xseries.XSound;
-import org.bukkit.potion.PotionEffect;
-import org.bukkit.potion.PotionEffectType;
-
-import java.util.Objects;
 
 /**
  * 游戏开始事件
- * 负责处理游戏开始时的初始化和团队升级效果
+ * 负责处理游戏开始时的初始化
  *
  * @author an5w1r@163.com
  */
@@ -35,7 +25,7 @@ public class GameStartEvent extends AbstractGameEvent {
      * 创建游戏开始事件
      */
     public GameStartEvent() {
-        super(MESSAGE_EVENT.getEventName(), CONFIG.getCountDown(), CONFIG.getEventPriority());
+        super(MESSAGE_EVENT.getEventName(), PLUGIN.getEventConfig().getStartEvent().getCountDown(), PLUGIN.getEventConfig().getStartEvent().getEventPriority());
     }
 
     /**
@@ -56,113 +46,15 @@ public class GameStartEvent extends AbstractGameEvent {
 
     /**
      * 执行游戏开始事件
-     * 注册团队升级任务、启动资源生成和指南针追踪
+     * 启动资源生成和指南针追踪
      *
      * @param gameManager 游戏实例
      */
     @Override
     public void execute(GameManager gameManager) {
-        registerTeamUpgradeTask(gameManager);
         startResourceGenerators(gameManager);
         startCompassTracking();
     }
-
-    /**
-     * 注册团队升级任务，处理团队效果和陷阱
-     *
-     * @param gameManager 游戏实例
-     */
-    private void registerTeamUpgradeTask(GameManager gameManager) {
-        gameManager.getGameEventManager().registerRunnable(MESSAGE_EVENT.getTeamUpgradeTaskName(), (s, c) ->
-            GamePlayer.getOnlinePlayers().forEach(player -> {
-                if (player.isSpectator()) {
-                    return;
-                }
-
-                for (GameTeam gameTeam : gameManager.getGameTeams()) {
-                    // 跳过不同世界的团队
-                    if (!Objects.equals(player.getPlayer().getLocation().getWorld(), gameTeam.getSpawnLocation().getWorld())) {
-                        continue;
-                    }
-
-                    // 处理疯狂矿工效果
-                    applyManicMinerEffect(gameTeam);
-
-                    // 如果是团队成员，检查治愈池效果
-                    if (gameTeam.isInTeam(player)) {
-                        applyHealingPoolEffect(player, gameTeam);
-                        continue;
-                    }
-
-                    // 处理敌方玩家进入团队领地的陷阱触发
-                    handleEnemyInTeamTerritory(player, gameTeam);
-                }
-            })
-        );
-    }
-
-    /**
-     * 应用疯狂矿工效果给团队成员
-     *
-     * @param gameTeam 游戏团队
-     */
-    private void applyManicMinerEffect(GameTeam gameTeam) {
-        if (gameTeam.getMagicMinerUpgrade() <= 0) {
-            return;
-        }
-
-        AzuraBedWars.getInstance().mainThreadRunnable(() -> gameTeam.getAlivePlayers().forEach((player -> {
-            PotionEffectType fastDigging = XPotion.HASTE.get();
-            if (fastDigging != null) {
-                // Subtract 1 from the level because Minecraft potion effect levels start at 0
-                int effectLevel = gameTeam.getMagicMinerUpgrade() - 1;
-                player.getPlayer().addPotionEffect(new PotionEffect(fastDigging,
-                    CONFIG.getUpgrade().getHasteEffectDuration(),
-                        effectLevel));
-            }
-        })));
-    }
-
-    /**
-     * 为团队基地内的玩家应用治愈池效果
-     *
-     * @param gamePlayer 游戏玩家
-     * @param gameTeam 游戏团队
-     */
-    private void applyHealingPoolEffect(GamePlayer gamePlayer, GameTeam gameTeam) {
-        double distance = gamePlayer.getPlayer().getLocation().distance(gameTeam.getSpawnLocation());
-
-        if (distance <= CONFIG.getUpgrade().getHealingPoolRange() && gameTeam.isHasHealPoolUpgrade()) {
-            AzuraBedWars.getInstance().mainThreadRunnable(() -> {
-                PotionEffectType regeneration = XPotion.REGENERATION.get();
-                if (regeneration != null) {
-                    gamePlayer.getPlayer().addPotionEffect(new PotionEffect(regeneration,
-                        CONFIG.getUpgrade().getRegenerationEffectDuration(),
-                        CONFIG.getUpgrade().getRegenerationEffectAmplifier()));
-                }
-            });
-        }
-    }
-
-    /**
-     * 处理敌方玩家进入团队领地触发的陷阱
-     *
-     * @param gamePlayer 游戏玩家
-     * @param gameTeam 游戏团队
-     */
-    private void handleEnemyInTeamTerritory(GamePlayer gamePlayer, GameTeam gameTeam) {
-        double distance = gamePlayer.getPlayer().getLocation().distance(gameTeam.getSpawnLocation());
-        TrapManager trapManager = gameTeam.getTrapManager();
-
-        if (distance <= CONFIG.getUpgrade().getTrapTriggerRange() && !gameTeam.isDead()) {
-            // 直接遍历激活的陷阱，使用策略模式触发
-            for (TrapType trapType : trapManager.getActiveTraps()) {
-                LoggerUtil.debug("GameStartEvent$handleEnemyInTeamTerritory | trigger " + trapType.name() + " trap, gamePlayer: " + gamePlayer.getName());
-                trapManager.getTrapStrategy(trapType).triggerTrap(gamePlayer, gameTeam);
-            }
-        }
-    }
-
 
 
     /**
